refactor m3u fetching
