+++ packages/media-parser/src/containers/m3u/run-over-m3u.ts
@@ -1,8 +1,7 @@
 import type {LogLevel} from '../../log';
 import {Log} from '../../log';
-import {registerAudioTrack, registerVideoTrack} from '../../register-track';
 import type {ParserState} from '../../state/parser-state';
-import {iteratorOverSegmentFiles} from './iterate-over-segment-files';
+import {processM3uChunk} from './process-m3u-chunk';
 import type {M3uStructure} from './types';
 
 export const runOverM3u = async ({
@@ -45,100 +44,12 @@ export const runOverM3u = async ({
 	}
 
 	Log.trace(logLevel, 'Starting new M3U parsing process for', playlistUrl);
-	return new Promise<void>((resolve, reject) => {
-		const run = iteratorOverSegmentFiles({
-			playlistUrl,
-			structure,
-			onInitialProgress: (newRun) => {
-				state.m3u.setM3uStreamRun(playlistUrl, newRun);
-				resolve();
-			},
-			logLevel: state.logLevel,
-			onDoneWithTracks() {
-				const allDone = state.m3u.setTracksDone(playlistUrl);
-				if (allDone) {
-					state.callbacks.tracks.setIsDone(state.logLevel);
-				}
-			},
-			onAudioTrack: audioDone
-				? null
-				: async (track) => {
-						const existingTracks = state.callbacks.tracks.getTracks();
-						let {trackId} = track;
-						while (existingTracks.find((t) => t.trackId === trackId)) {
-							trackId++;
-						}
 
-						const onAudioSample = await registerAudioTrack({
-							container: 'm3u8',
-							track: {
-								...track,
-								trackId,
-							},
-							registerAudioSampleCallback:
-								state.callbacks.registerAudioSampleCallback,
-							tracks: state.callbacks.tracks,
-							logLevel: state.logLevel,
-							onAudioTrack: state.onAudioTrack,
-						});
-						state.m3u.sampleSorter.addToStreamWithTrack(playlistUrl);
-
-						if (onAudioSample === null) {
-							return null;
-						}
-
-						state.m3u.sampleSorter.addAudioStreamToConsider(
-							playlistUrl,
-							onAudioSample,
-						);
-
-						return async (sample) => {
-							await state.m3u.sampleSorter.addAudioSample(playlistUrl, sample);
-						};
-					},
-			onVideoTrack: videoDone
-				? null
-				: async (track) => {
-						const existingTracks = state.callbacks.tracks.getTracks();
-						let {trackId} = track;
-						while (existingTracks.find((t) => t.trackId === trackId)) {
-							trackId++;
-						}
-
-						const onVideoSample = await registerVideoTrack({
-							container: 'm3u8',
-							track: {
-								...track,
-								trackId,
-							},
-							logLevel: state.logLevel,
-							onVideoTrack: state.onVideoTrack,
-							registerVideoSampleCallback:
-								state.callbacks.registerVideoSampleCallback,
-							tracks: state.callbacks.tracks,
-						});
-						state.m3u.sampleSorter.addToStreamWithTrack(playlistUrl);
-
-						if (onVideoSample === null) {
-							return null;
-						}
-
-						state.m3u.sampleSorter.addVideoStreamToConsider(
-							playlistUrl,
-							onVideoSample,
-						);
-
-						return async (sample) => {
-							await state.m3u.sampleSorter.addVideoSample(playlistUrl, sample);
-						};
-					},
-			m3uState: state.m3u,
-			parentController: state.controller,
-			readerInterface: state.readerInterface,
-		});
-
-		run.catch((err) => {
-			reject(err);
-		});
+	await processM3uChunk({
+		playlistUrl,
+		state,
+		structure,
+		audioDone,
+		videoDone,
 	});
 };
